+++ HomeControlSystem/2_CoreSystem/HardwarePins/mHardwarePins.h
@@ -35,7 +35,7 @@ class mHardwarePins :
     static const char* PM_MODULE_CORE_HARDWAREPINS_FRIENDLY_CTR;
     PGM_P GetModuleName(){          return PM_MODULE_CORE_HARDWAREPINS_CTR; }
     PGM_P GetModuleFriendlyName(){  return PM_MODULE_CORE_HARDWAREPINS_FRIENDLY_CTR; }
-    uint8_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_CORE_HARDWAREPINS_ID; }
+    uint16_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_CORE_HARDWAREPINS_ID; }
 
     #ifdef USE_DEBUG_CLASS_SIZE
     uint16_t GetClassSize(){
