+++ HomeControlSystem/2_CoreSystem/RuleEngine/mRuleEngine.h
@@ -36,7 +36,7 @@ class mRuleEngine :
     static const char* PM_MODULE_CORE_RULES_FRIENDLY_CTR;
     PGM_P GetModuleName(){          return PM_MODULE_CORE_RULES_CTR; }
     PGM_P GetModuleFriendlyName(){  return PM_MODULE_CORE_RULES_FRIENDLY_CTR; }
-    uint8_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_CORE_RULES_FRIENDLY_ID; }
+    uint16_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_CORE_RULES_FRIENDLY_ID; }
 
     #ifdef USE_DEBUG_CLASS_SIZE
     uint16_t GetClassSize(){
