+++ HomeControlSystem/2_CoreSystem/Updates/mUpdates.h
@@ -24,7 +24,7 @@ class mUpdates :
     static const char* PM_MODULE_CORE_UPDATES_FRIENDLY_CTR;
     PGM_P GetModuleName(){         return PM_MODULE_CORE_UPDATES_CTR; }
     PGM_P GetModuleFriendlyName(){ return PM_MODULE_CORE_UPDATES_FRIENDLY_CTR; }
-    uint8_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_CORE_UPDATES_ID; }
+    uint16_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_CORE_UPDATES_ID; }
 
     
     #ifdef USE_DEBUG_CLASS_SIZE
