+++ HomeControlSystem/3_Network/WebServer/mWebServer.h
@@ -1298,7 +1298,7 @@ public mTaskerInterface{
     static const char* PM_MODULE_NETWORK_WEBSERVER_FRIENDLY_CTR;
     PGM_P GetModuleName(){          return PM_MODULE_NETWORK_WEBSERVER_CTR; }
     PGM_P GetModuleFriendlyName(){  return PM_MODULE_NETWORK_WEBSERVER_FRIENDLY_CTR; }
-    uint8_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_NETWORK_WEBSERVER_ID; }
+    uint16_t GetModuleUniqueID(){ return D_UNIQUE_MODULE_NETWORK_WEBSERVER_ID; }
 
     #ifdef USE_DEBUG_CLASS_SIZE
     uint16_t GetClassSize(){
